requirements.txt
pynput==1.7.1
pyserial==3.5
